scripts/requirements.txt
resend==0.8.0
python-dotenv==1.0.0 